requirements.txt
Django==1.6.5
dj-database-url==0.3.0
dj-static==0.0.5
django-haystack==2.1.0
gunicorn==19.0.0
honcho==0.5.0
pyelasticsearch==0.6.1
pystache==0.5.4
requests==2.3.0
simplejson==3.5.3
six==1.7.3
static==1.0.2
wsgiref==0.1.2
